Mergetentngs

+++ src/scalarmappedfe.cpp
@@ -1,6 +1,4 @@
 #include "scalarmappedfe.hpp"
-#include "h1lofe.hpp"
-#include "l2hofe.hpp"
 
 namespace ngfem
 {
@@ -51,7 +49,7 @@ namespace ngfem
   BaseScalarMappedElement ::CalcShape (const BaseMappedIntegrationRule &mir,
                                        SliceMatrix<> shape) const
   {
-    for (int i = 0; i < mir.Size (); i++)
+    for (size_t i = 0; i < mir.Size (); i++)
       CalcShape (mir[i], shape.Col (i));
   }
 
@@ -118,7 +116,7 @@ namespace ngfem
   {
     VectorMem<20, double> shape (ndof);
     coefs.Range (0, ndof) = 0.0;
-    for (int i = 0; i < mir.Size (); i++) // GetNIP()
+    for (size_t i = 0; i < mir.Size (); i++) // GetNIP()
       {
         CalcShape (mir[i], shape);
         coefs.Range (0, ndof) += vals (i) * shape;
@@ -139,7 +137,7 @@ namespace ngfem
       const SIMD_BaseMappedIntegrationRule &mir,
       BareSliceMatrix<SIMD<double>> values, SliceMatrix<> coefs) const
   {
-    for (int i = 0; i < coefs.Width (); i++)
+    for (size_t i = 0; i < coefs.Width (); i++)
       AddTrans (mir, values.Row (i), coefs.Col (i));
   }
 
@@ -187,12 +185,12 @@ namespace ngfem
     sparsemat[0].SetSize (0);
     sparsemat[1].SetSize (0);
     sparsemat[2].SetSize (0);
-    for (int i = 0; i < mat.Height (); i++)
+    for (size_t i = 0; i < mat.Height (); i++)
       {
         // guarantee sparsemat[0].Size() to be nbasis,
         // could be shorter but easier for mat*vec
         sparsemat[0].Append (spsize);
-        for (int j = 0; j < mat.Width (); j++)
+        for (size_t j = 0; j < mat.Width (); j++)
           {
             if (mat (i, j))
               {
@@ -215,7 +213,7 @@ namespace ngfem
   ScalarMappedElement<D>::CalcDShape (const BaseMappedIntegrationRule &mir,
                                       BareSliceMatrix<> dshapes) const
   {
-    for (int i = 0; i < mir.Size (); i++)
+    for (size_t i = 0; i < mir.Size (); i++)
       CalcDShape (mir[i], dshapes.Cols (i * D, (i + 1) * D));
   }
 
@@ -232,7 +230,7 @@ namespace ngfem
   void ScalarMappedElement<D>::CalcMappedDShape (
       const BaseMappedIntegrationRule &mir, SliceMatrix<> dshapes) const
   {
-    for (int i = 0; i < mir.Size (); i++)
+    for (size_t i = 0; i < mir.Size (); i++)
       CalcMappedDShape (mir[i], dshapes.Cols (i * D, (i + 1) * D));
   }
 
@@ -263,7 +261,7 @@ namespace ngfem
   {
     MatrixFixWidth<D> dshape (ndof);
     coefs.Range (0, ndof) = 0.0;
-    for (int i = 0; i < ir.Size (); i++)
+    for (size_t i = 0; i < ir.Size (); i++)
       {
         CalcDShape (ir[i], dshape);
         coefs.Range (0, ndof) += dshape * vals.Row (i);
@@ -366,7 +364,7 @@ namespace ngfem
     STACK_ARRAY (double, mem2, 2 * (order + 1) + 1);
     mem2[0] = 0;
     double *polxt[2];
-    for (size_t d = 0; d < 2; d++)
+    for (int d = 0; d < 2; d++)
       {
         polxt[d] = &mem2[d * (order + 1) + 1];
         Monomial (order, cpoint[d], polxt[d]);
@@ -377,8 +375,8 @@ namespace ngfem
         for (int d2 = 0; d2 < 2; d2++)
           {
             Vector<double> pol (npoly);
-            for (size_t i = 0, ii = 0; i <= order; i++)
-              for (size_t j = 0; j <= order - i; j++)
+            for (int i = 0, ii = 0; i <= order; i++)
+              for (int j = 0; j <= order - i; j++)
                 pol[ii++] = (d1 != d2 ? i * j
                                       : (d1 == 0 ? i * (i - 1) : j * (j - 1)))
                             * polxt[0][i - (d1 == 0) - (d2 == 0)]
@@ -410,7 +408,7 @@ namespace ngfem
     STACK_ARRAY (double, mem2, 3 * (order + 1) + 1);
     mem2[0] = 0;
     double *polxt[3];
-    for (size_t d = 0; d < 3; d++)
+    for (int d = 0; d < 3; d++)
       {
         polxt[d] = &mem2[d * (order + 1) + 1];
         Monomial (order, cpoint[d], polxt[d]);
@@ -469,7 +467,7 @@ namespace ngfem
       const SIMD_BaseMappedIntegrationRule &smir,
       BareSliceMatrix<SIMD<double>> shape) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<2, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -478,15 +476,15 @@ namespace ngfem
         // calc 1 dimensional monomial basis
         STACK_ARRAY (SIMD<double>, mem, 2 * (order + 1));
         Vec<2, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 2; d++)
+        for (int d = 0; d < 2; d++)
           {
             polxt[d] = &mem[d * (order + 1)];
             Monomial (order, cpoint[d], polxt[d]);
           }
         // calc D+1 dimenional monomial basis
         Vector<SIMD<double>> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
             pol[ii++] = polxt[0][i] * polxt[1][j];
         // TB*monomials for trefftz shape fcts
         for (int i = 0; i < this->ndof; ++i)
@@ -503,7 +501,7 @@ namespace ngfem
       const SIMD_BaseMappedIntegrationRule &smir,
       BareSliceMatrix<SIMD<double>> shape) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<3, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -512,16 +510,16 @@ namespace ngfem
         // calc 1 dimensional monomial basis
         STACK_ARRAY (SIMD<double>, mem, 3 * (order + 1));
         Vec<3, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 3; d++)
+        for (int d = 0; d < 3; d++)
           {
             polxt[d] = &mem[d * (order + 1)];
             Monomial (order, cpoint[d], polxt[d]);
           }
         // calc D+1 dimenional monomial basis
         Vector<SIMD<double>> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
-            for (size_t k = 0; k <= order - i - j; k++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
               pol[ii++] = polxt[0][i] * polxt[1][j] * polxt[2][k];
         // TB*monomials for trefftz shape fcts
         for (int i = 0; i < this->ndof; ++i)
@@ -540,7 +538,7 @@ namespace ngfem
   {
     // auto & smir = static_cast<const SIMD_MappedIntegrationRule<D-1,D>&>
     // (mir);
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<4, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -549,17 +547,17 @@ namespace ngfem
         // calc 1 dimensional monomial basis
         STACK_ARRAY (SIMD<double>, mem, 4 * (order + 1));
         Vec<4, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 4; d++)
+        for (int d = 0; d < 4; d++)
           {
             polxt[d] = &mem[d * (order + 1)];
             Monomial (order, cpoint[d], polxt[d]);
           }
         // calc D+1 dimenional monomial basis
         Vector<SIMD<double>> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
-            for (size_t k = 0; k <= order - i - j; k++)
-              for (size_t l = 0; l <= order - i - j - k; l++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
+              for (int l = 0; l <= order - i - j - k; l++)
                 pol[ii++]
                     = polxt[0][i] * polxt[1][j] * polxt[2][k] * polxt[3][l];
         // TB*monomials for trefftz shape fcts
@@ -585,7 +583,7 @@ namespace ngfem
       const SIMD_BaseMappedIntegrationRule &smir,
       BareSliceMatrix<SIMD<double>> dshape) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<2, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -596,7 +594,7 @@ namespace ngfem
         STACK_ARRAY (SIMD<double>, mem, 2 * (order + 1) + 1);
         mem[0] = 0;
         Vec<2, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 2; d++)
+        for (int d = 0; d < 2; d++)
           {
             polxt[d] = &mem[d * (order + 1) + 1];
             Monomial (order, cpoint[d], polxt[d]);
@@ -605,8 +603,8 @@ namespace ngfem
         for (int d = 0; d < 2; d++)
           {
             Vector<SIMD<double>> pol (npoly);
-            for (size_t i = 0, ii = 0; i <= order; i++)
-              for (size_t j = 0; j <= order - i; j++)
+            for (int i = 0, ii = 0; i <= order; i++)
+              for (int j = 0; j <= order - i; j++)
                 pol[ii++] = (d == 0 ? i : (d == 1 ? j : 0))
                             * polxt[0][i - (d == 0)] * polxt[1][j - (d == 1)];
 
@@ -628,7 +626,7 @@ namespace ngfem
       const SIMD_BaseMappedIntegrationRule &smir,
       BareSliceMatrix<SIMD<double>> dshape) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<3, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -639,7 +637,7 @@ namespace ngfem
         STACK_ARRAY (SIMD<double>, mem, 3 * (order + 1) + 1);
         mem[0] = 0;
         Vec<3, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 3; d++)
+        for (int d = 0; d < 3; d++)
           {
             polxt[d] = &mem[d * (order + 1) + 1];
             Monomial (order, cpoint[d], polxt[d]);
@@ -648,9 +646,9 @@ namespace ngfem
         for (int d = 0; d < 3; d++)
           {
             Vector<SIMD<double>> pol (npoly);
-            for (size_t i = 0, ii = 0; i <= order; i++)
-              for (size_t j = 0; j <= order - i; j++)
-                for (size_t k = 0; k <= order - i - j; k++)
+            for (int i = 0, ii = 0; i <= order; i++)
+              for (int j = 0; j <= order - i; j++)
+                for (int k = 0; k <= order - i - j; k++)
                   pol[ii++] = (d == 0 ? i : (d == 1 ? j : (d == 2 ? k : 0)))
                               * polxt[0][i - (d == 0)] * polxt[1][j - (d == 1)]
                               * polxt[2][k - (d == 2)];
@@ -673,7 +671,7 @@ namespace ngfem
       const SIMD_BaseMappedIntegrationRule &smir,
       BareSliceMatrix<SIMD<double>> dshape) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<4, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -684,7 +682,7 @@ namespace ngfem
         STACK_ARRAY (SIMD<double>, mem, 4 * (order + 1) + 1);
         mem[0] = 0;
         Vec<4, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 4; d++)
+        for (int d = 0; d < 4; d++)
           {
             polxt[d] = &mem[d * (order + 1) + 1];
             Monomial (order, cpoint[d], polxt[d]);
@@ -693,10 +691,10 @@ namespace ngfem
         for (int d = 0; d < 4; d++)
           {
             Vector<SIMD<double>> pol (npoly);
-            for (size_t i = 0, ii = 0; i <= order; i++)
-              for (size_t j = 0; j <= order - i; j++)
-                for (size_t k = 0; k <= order - i - j; k++)
-                  for (size_t l = 0; l <= order - i - j - k; l++)
+            for (int i = 0, ii = 0; i <= order; i++)
+              for (int j = 0; j <= order - i; j++)
+                for (int k = 0; k <= order - i - j; k++)
+                  for (int l = 0; l <= order - i - j - k; l++)
                     pol[ii++]
                         = (d == 0 ? i
                                   : (d == 1 ? j
@@ -742,15 +740,15 @@ namespace ngfem
     // calc 1 dimensional monomial basis
     STACK_ARRAY (double, mem, 2 * (order + 1));
     double *polxt[2];
-    for (size_t d = 0; d < 2; d++)
+    for (int d = 0; d < 2; d++)
       {
         polxt[d] = &mem[d * (order + 1)];
         Monomial (order, cpoint[d], polxt[d]);
       }
     // calc D+1 dimenional monomial basis
     Vector<double> pol (npoly);
-    for (size_t i = 0, ii = 0; i <= order; i++)
-      for (size_t j = 0; j <= order - i; j++)
+    for (int i = 0, ii = 0; i <= order; i++)
+      for (int j = 0; j <= order - i; j++)
         pol[ii++] = polxt[0][i] * polxt[1][j];
     // TB*monomials for trefftz shape fcts
     for (int i = 0; i < this->ndof; ++i)
@@ -774,16 +772,16 @@ namespace ngfem
     STACK_ARRAY (double, mem, 3 * (order + 1));
     // double* polxt[3];
     Vec<3, double *> polxt;
-    for (size_t d = 0; d < 3; d++)
+    for (int d = 0; d < 3; d++)
       {
         polxt[d] = &mem[d * (order + 1)];
         Monomial (order, cpoint[d], polxt[d]);
       }
     // calc D+1 dimenional monomial basis
     Vector<double> pol (npoly);
-    for (size_t i = 0, ii = 0; i <= order; i++)
-      for (size_t j = 0; j <= order - i; j++)
-        for (size_t k = 0; k <= order - i - j; k++)
+    for (int i = 0, ii = 0; i <= order; i++)
+      for (int j = 0; j <= order - i; j++)
+        for (int k = 0; k <= order - i - j; k++)
           pol[ii++] = polxt[0][i] * polxt[1][j] * polxt[2][k];
     // TB*monomials for trefftz shape fcts
     for (int i = 0; i < this->ndof; ++i)
@@ -806,17 +804,17 @@ namespace ngfem
     // calc 1 dimensional monomial basis
     STACK_ARRAY (double, mem, 4 * (order + 1));
     double *polxt[4];
-    for (size_t d = 0; d < 4; d++)
+    for (int d = 0; d < 4; d++)
       {
         polxt[d] = &mem[d * (order + 1)];
         Monomial (order, cpoint[d], polxt[d]);
       }
     // calc D+1 dimenional monomial basis
     Vector<double> pol (npoly);
-    for (size_t i = 0, ii = 0; i <= order; i++)
-      for (size_t j = 0; j <= order - i; j++)
-        for (size_t k = 0; k <= order - i - j; k++)
-          for (size_t l = 0; l <= order - i - j - k; l++)
+    for (int i = 0, ii = 0; i <= order; i++)
+      for (int j = 0; j <= order - i; j++)
+        for (int k = 0; k <= order - i - j; k++)
+          for (int l = 0; l <= order - i - j - k; l++)
             pol[ii++] = polxt[0][i] * polxt[1][j] * polxt[2][k] * polxt[3][l];
     // TB*monomials for trefftz shape fcts
     for (int i = 0; i < this->ndof; ++i)
@@ -848,7 +846,7 @@ namespace ngfem
     STACK_ARRAY (double, mem, 2 * (order + 1) + 1);
     mem[0] = 0;
     double *polxt[2];
-    for (size_t d = 0; d < 2; d++)
+    for (int d = 0; d < 2; d++)
       {
         polxt[d] = &mem[d * (order + 1) + 1];
         Monomial (order, cpoint[d], polxt[d]);
@@ -857,8 +855,8 @@ namespace ngfem
     for (int d = 0; d < 2; d++)
       {
         Vector<double> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
             pol[ii++] = (d == 0 ? i : (d == 1 ? j : 0))
                         * polxt[0][i - (d == 0)] * polxt[1][j - (d == 1)];
 
@@ -888,7 +886,7 @@ namespace ngfem
     mem[0] = 0;
     // double* polxt[4];
     Vec<3, double *> polxt;
-    for (size_t d = 0; d < 3; d++)
+    for (int d = 0; d < 3; d++)
       {
         polxt[d] = &mem[d * (order + 1) + 1];
         Monomial (order, cpoint[d], polxt[d]);
@@ -897,9 +895,9 @@ namespace ngfem
     for (int d = 0; d < 3; d++)
       {
         Vector<double> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
-            for (size_t k = 0; k <= order - i - j; k++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
               pol[ii++] = (d == 0 ? i : (d == 1 ? j : (d == 2 ? k : 0)))
                           * polxt[0][i - (d == 0)] * polxt[1][j - (d == 1)]
                           * polxt[2][k - (d == 2)];
@@ -929,7 +927,7 @@ namespace ngfem
     STACK_ARRAY (double, mem, 4 * (order + 1) + 1);
     mem[0] = 0;
     double *polxt[4];
-    for (size_t d = 0; d < 4; d++)
+    for (int d = 0; d < 4; d++)
       {
         polxt[d] = &mem[d * (order + 1) + 1];
         Monomial (order, cpoint[d], polxt[d]);
@@ -938,10 +936,10 @@ namespace ngfem
     for (int d = 0; d < 4; d++)
       {
         Vector<double> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
-            for (size_t k = 0; k <= order - i - j; k++)
-              for (size_t l = 0; l <= order - i - j - k; l++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
+              for (int l = 0; l <= order - i - j - k; l++)
                 pol[ii++]
                     = (d == 0 ? i
                               : (d == 1 ? j : (d == 2 ? k : (d == 3 ? l : 0))))
@@ -986,7 +984,7 @@ namespace ngfem
       BareSliceMatrix<SIMD<double>> wavespeed,
       BareSliceMatrix<SIMD<double>> mu) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<2, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -996,15 +994,15 @@ namespace ngfem
         mem[0] = 0;
         mem[1] = 0;
         Vec<2, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 2; d++)
+        for (int d = 0; d < 2; d++)
           {
             polxt[d] = &mem[d * (order + 1) + 2];
             Monomial (order, cpoint[d], polxt[d]);
           }
 
         Vector<SIMD<double>> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
             pol[ii++] = (i * (i - 1) * polxt[0][i - 2] * polxt[1][j]
                          - j * (j - 1) * polxt[0][i] * polxt[1][j - 2]
                                * wavespeed (0, imip))
@@ -1029,7 +1027,7 @@ namespace ngfem
       BareSliceMatrix<SIMD<double>> wavespeed,
       BareSliceMatrix<SIMD<double>> mu) const
   {
-    for (int imip = 0; imip < smir.Size (); imip++)
+    for (size_t imip = 0; imip < smir.Size (); imip++)
       {
         Vec<3, SIMD<double>> cpoint = smir[imip].GetPoint ();
         cpoint -= elcenter;
@@ -1039,16 +1037,16 @@ namespace ngfem
         mem[0] = 0;
         mem[1] = 0;
         Vec<3, SIMD<double> *> polxt;
-        for (size_t d = 0; d < 3; d++)
+        for (int d = 0; d < 3; d++)
           {
             polxt[d] = &mem[d * (order + 1) + 2];
             Monomial (order, cpoint[d], polxt[d]);
           }
 
         Vector<SIMD<double>> pol (npoly);
-        for (size_t i = 0, ii = 0; i <= order; i++)
-          for (size_t j = 0; j <= order - i; j++)
-            for (size_t k = 0; k <= order - i - j; k++)
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
               pol[ii++]
                   = (i * (i - 1) * polxt[0][i - 2] * polxt[1][j] * polxt[2][k]
                      + j * (j - 1) * polxt[0][i] * polxt[1][j - 2]
